scripts/eco_validate/requirements.txt
matplotlib==3.5.2
pandas==1.3.5
PyYAML==6.0
seaborn==0.11.2
setuptools==58.0.4
